+++ app/src/js/components/BatchAsyncCommands/BatchAsyncCommands.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect, useState, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { withRouter } from 'react-router';
 import queue from 'stubborn-queue';
@@ -32,201 +32,182 @@ const CONCURRENCY = 3;
  * All those scenarios can display different content for the modal based on logic setup
  * within getModalOptions.
  */
+export const BatchCommand = ({
+  action,
+  dispatch,
+  state,
+  text,
+  selected,
+  className,
+  clearError,
+  onSuccess,
+  onError,
+  confirm,
+  confirmOptions,
+  getModalOptions,
+  history,
+}) => {
+  const [callbacks, setCallbacks] = useState({});
+  const [activeModal, setActiveModal] = useState(false);
+  const [completed, setCompleted] = useState(0);
+  const [status, setStatus] = useState(null);
+  const [modalOpts, setModalOpts] = useState(null);
+  const [errorMsg, setErrorMsg] = useState(null);
+  const [meta, setMeta] = useState({});
+  const [results, setResults] = useState(null);
+  const isRunning = useRef(false);
 
-export class BatchCommand extends React.Component {
-  constructor() {
-    super();
-    this.state = {
-      callbacks: {},
-      activeModal: false,
-      completed: 0,
-      status: null,
-      modalOptions: null,
-      errorMessage: null,
-      meta: {},
+  const buildId = (item) => (typeof item === 'string' ? item : item.granuleId);
+
+  const updateMeta = (newMeta) => {
+    setMeta((prevMeta) => ({ ...prevMeta, ...newMeta }));
+  };
+
+  const closeModal = () => {
+    setActiveModal(false);
+    setMeta({});
+  };
+
+  useEffect(() => {
+    if (isRunning.current) return;
+    isRunning.current = true;
+    const handleCallbacks = () => {
+      // on success or error, call and remove the saved callback
+      Object.keys(callbacks).forEach((id) => {
+        if (!state[id] || !callbacks[id]) return;
+        if (state[id].status === 'success') callbacks[id](null, id);
+        else if (state[id].status === 'error') { callbacks[id]({ error: state[id].error, id }); }
+
+        if (state[id].status === 'success' || state[id].status === 'error') {
+          const newCallbacks = { ...callbacks };
+          delete newCallbacks[id];
+          setCallbacks(newCallbacks);
+          setCompleted((prevCompleted) => prevCompleted + 1);
+        }
+      });
+      isRunning.current = false;
     };
-    this.isRunning = false;
-    this.buildId = this.buildId.bind(this);
-    this.confirm = this.confirm.bind(this);
-    this.cancel = this.cancel.bind(this);
-    this.start = this.start.bind(this);
-    this.initAction = this.initAction.bind(this);
-    this.onComplete = this.onComplete.bind(this);
-    this.createErrorMessage = this.createErrorMessage.bind(this);
-    this.cleanup = this.cleanup.bind(this);
-    this.isInflight = this.isInflight.bind(this);
-    this.handleClick = this.handleClick.bind(this);
-    this.closeModal = this.closeModal.bind(this);
-    this.updateMeta = this.updateMeta.bind(this);
-  }
+    handleCallbacks();
+  }, [state, callbacks]);
 
-  buildId(item) {
-    return (typeof item === 'string') ? item : item.granuleId;
-  }
-
-  updateMeta(meta) {
-    this.setState({ meta: { ...this.state.meta, ...meta } });
-  }
-
-  closeModal() {
-    this.setState({ activeModal: false });
-    this.setState({ meta: {} });
-  }
-
-  componentDidUpdate() {
-    if (this.isRunning) return;
-    this.isRunning = true;
-    const { state } = this.props;
-    const { callbacks, completed } = this.state;
-
-    // on success or error, call and remove the saved callback
-    Object.keys(callbacks).forEach((id) => {
-      if (!state[id] || !callbacks[id]) return;
-      if (state[id].status === 'success') callbacks[id](null, id);
-      else if (state[id].status === 'error') { callbacks[id]({ error: state[id].error, id }); }
-
-      if (state[id].status === 'success' || state[id].status === 'error') {
-        delete callbacks[id];
-        this.setState({ callbacks, completed: completed + 1 });
-      }
-    });
-    this.isRunning = false;
-  }
-
-  confirm() {
-    const { selected, history, getModalOptions } = this.props;
+  const confirmAction = () => {
     if (typeof getModalOptions === 'function') {
       const modalOptions = getModalOptions({
         selected,
         history,
         isOnModalConfirm: true,
         isOnModalComplete: false,
-        closeModal: this.closeModal,
+        closeModal
       });
-      this.setState({ modalOptions });
+      setModalOpts(modalOptions);
 
       // if we're replacing the onConfirm function, we don't want to continue with the current one
       if (modalOptions.onConfirm) {
         return;
       }
     }
-    if (!this.isInflight()) this.start();
-  }
+    if (!isInflight()) start();
+  };
 
-  cancel() {
+  const cancel = () => {
     // prevent cancel when inflight
-    if (!this.isInflight()) this.setState({ activeModal: false });
-  }
+    if (!isInflight()) setActiveModal(false);
+  };
 
-  start() {
-    const { selected } = this.props;
+  const start = () => {
     // if we have inflight callbacks, don't allow further clicks
-    if (!Array.isArray(selected) || !selected.length || this.isInflight()) { return false; }
-    const q = queue(CONCURRENCY);
-    for (let i = 0; i < selected.length; i += 1) {
-      q.add(this.initAction, this.buildId(selected[i]));
-    }
-    q.done(this.onComplete);
-  }
+    if (!Array.isArray(selected) || !selected.length || isInflight()) { return false; }
+    const q = queue(initAction, CONCURRENCY);
+    selected.forEach((item) => {
+      q.add(initAction, buildId(item));
+    });
+    q.done(onComplete);
+  };
 
   // save a reference to the callback in state, then init the action
-  initAction(id, callback) {
-    const { dispatch, action } = this.props;
-    const { callbacks, meta } = this.state;
-    callbacks[id] = callback;
-    this.setState({ callbacks });
+  const initAction = (id, callback) => {
+    setCallbacks((prevCallbacks) => ({
+      ...prevCallbacks,
+      [id]: callback
+    }));
     return dispatch(action(id, meta));
-  }
+  };
 
   // immediately change the UI to show either success or error
-  onComplete(errors, results) {
-    const {
-      getModalOptions,
-      selected,
-      history,
-    } = this.props;
+  const onComplete = (errors) => {
     // turn array of errors from queue into single error for ui
-    const errorMessage = this.createErrorMessage(errors);
-    this.setState({ errorMessage, results, status: errorMessage ? 'error' : 'success' });
+    const errMsg = createErrorMessage(errors);
+    setErrorMsg(errMsg);
+    setStatus(errMsg ? 'error' : 'success');
+    setResults(errors ? [] : selected.map((item) => buildId(item)));
     if (typeof getModalOptions === 'function') {
-      // setTimeout(() => {
       const modalOptions = getModalOptions({
         history,
         selected,
-        results,
+        results: selected.map((item) => buildId(item)),
         errors,
-        errorMessage,
+        errorMessage: errMsg,
         isOnModalComplete: true,
-        closeModal: this.closeModal,
+        closeModal
       });
-      this.setState({ modalOptions, status: null });
+      setModalOpts(modalOptions);
+      setStatus(null);
     }
-  }
+  };
 
   // combine multiple errors into one
-  createErrorMessage(errors) {
+  const createErrorMessage = (errors) => {
     if (!errors || !errors.length) return;
     return `${errors.length} error(s) occurred: \n${errors
       .map((err) => err.error.toString())
       .join('\n')}`;
-  }
+  };
 
   // call onSuccess and onError functions as needed
-  cleanup(e) {
-    const { errorMessage, results } = this.state;
-    const {
-      clearError,
-      dispatch,
-      onSuccess,
-      onError,
-      selected,
-    } = this.props;
-    if (errorMessage && typeof onError === 'function') onError(errorMessage);
-    if (results && results.length && typeof onSuccess === 'function') { onSuccess(results, errorMessage); }
-
+  const cleanup = () => {
+    if (errorMsg && typeof onError === 'function') onError(errorMsg);
+    if (results && results.length && typeof onSuccess === 'function') { onSuccess(results, errorMsg); }
     if (typeof clearError === 'function') {
-      selected.forEach((item) => dispatch(clearError(this.buildId(item))));
+      selected.forEach((item) => dispatch(clearError(buildId(item))));
     }
+    setActiveModal(false);
+    setCompleted(0);
+    setErrorMsg(null);
+    setStatus(null);
+    setResults(null);
+  };
 
-    this.setState({ activeModal: false, completed: 0, errorMessage: null, results: null, status: null });
-  }
+  const isInflight = () => !!Object.keys(callbacks).length;
 
-  isInflight() {
-    return !!Object.keys(this.state.callbacks).length;
-  }
-
-  handleClick() {
-    const { selected, history, getModalOptions } = this.props;
+  const handleClick = () => {
     if (typeof getModalOptions === 'function') {
       const modalOptions = getModalOptions({
-        onChange: this.updateMeta,
+        onChange: updateMeta,
         selected,
         history,
-        closeModal: this.closeModal,
+        closeModal
       });
-      this.setState({ modalOptions });
+      setModalOpts(modalOptions);
     }
-    if (this.props.confirm) {
-      this.setState({ activeModal: true, completed: 0 });
-    } else this.start();
-  }
+    if (confirm) {
+      setActiveModal(true);
+      setCompleted(0);
+    } else start();
+  };
 
-  render() {
-    const { text, selected, className, confirm, confirmOptions } = this.props;
-    const { activeModal, completed, errorMessage, status, modalOptions } = this.state;
-    const todo = selected.length;
-    const inflight = this.isInflight();
+  const todo = selected.length;
+  const inflight = isInflight();
 
-    // show button as disabled when loading, and in the delay before we clean up.
-    const buttonClass = inflight ? 'button--disabled' : '';
-    const confirmResult = confirm(todo);
-    const confirmTextArray = isArray(confirmResult) ? confirmResult : [confirmResult];
-    const percentage = todo ? ((completed * 100) / todo).toFixed(2) : 0;
+  // show button as disabled when loading, and in the delay before we clean up.
+  const buttonClass = inflight ? 'button--disabled' : '';
+  const confirmResult = confirm(todo);
+  const confirmTextArray = isArray(confirmResult) ? confirmResult : [confirmResult];
+  const percentage = todo ? ((completed * 100) / todo).toFixed(2) : 0;
 
-    return (
+  return (
       <div>
         <AsyncCommand
-          action={this.handleClick}
+          action={handleClick}
           text={text}
           className={className}
           hidden={!activeModal && (!todo || !!inflight)}
@@ -243,16 +224,16 @@ export class BatchCommand extends React.Component {
           <DefaultModal
             className="batch-async-modal"
             /* Need to separate cancel and close button functions because cancel is secondary and close is primary */
-            onCancel={status ? this.cleanup : this.cancel}
+            onCancel={status ? cleanup : cancel}
             cancelButtonText={status ? 'Close' : 'Cancel'}
-            onCloseModal={status ? this.cleanup : this.cancel}
-            onConfirm={this.confirm}
+            onCloseModal={status ? cleanup : cancel}
+            onConfirm={confirmAction}
             title={text}
             showModal={activeModal}
             confirmButtonClass={`${buttonClass} button--submit`}
             cancelButtonClass={`${buttonClass} button--cancel`}
             hasConfirmButton={!status} /* if status is set, we just want to close the modal */
-            {...modalOptions}
+            {...modalOpts}
             {...inflight ? { size: 'md' } : {}}
           >
             {inflight &&
@@ -273,11 +254,11 @@ export class BatchCommand extends React.Component {
             <>
               <div>
                 <Alert variant="danger"><strong>Error</strong></Alert>
-                <ErrorReport report= {errorMessage}/>
+                <ErrorReport report={errorMsg} />
               </div>
             </>
             }
-            {(!inflight && !status) && (!modalOptions || !modalOptions.children) && (
+            {(!inflight && !status) && (!modalOpts || !modalOpts.children) && (
               <>
                 <div>{confirmTextArray.map((confirmText, index) => (
                   <React.Fragment key={index}>
@@ -286,8 +267,8 @@ export class BatchCommand extends React.Component {
                 ))}</div>
                 <div className="modal__internal modal__formcenter">
                   {confirmOptions &&
-                    confirmOptions.map((option) => (
-                      <div key={`option-${confirmOptions.indexOf(option)}`}>
+                    confirmOptions.map((option, index) => (
+                      <div key={`option-${index}`}>
                         {option}
                         <br />
                       </div>))
@@ -295,13 +276,12 @@ export class BatchCommand extends React.Component {
                 </div>
               </>
             )}
-            {modalOptions && modalOptions.children}
+            {modalOpts && modalOpts.children}
           </DefaultModal>
         </div>
       </div>
-    );
-  }
-}
+  );
+};
 
 BatchCommand.propTypes = {
   action: PropTypes.func,