+++ CHANGELOG.md
@@ -23,6 +23,9 @@ This version of the dashboard requires Cumulus API >= v18.2.0
     `@cumulus/aws-client@18.2.0-alpha.0` to work with localstack v3.0.0
 - **Snyk Security Fix**
   - Upgraded axios from 1.6.0 to 1.6.8
+- **CUMULUS-2668**
+  - Refactored DeleteCollectionModal and BatchAsyncCommands from class to function components.
+  - Modified unit tests to accomodate new changes.
 - **CUMULUS-2356**
   - Removed gitc localization
   - Allowlisted braces, ws, and fast-xml-parser